Fix checkbox test
