+++ src/modules/frontend/listItems.js
@@ -1,5 +1,5 @@
 import {
-  dragstart, dragover, dragleave, drop, dragend, remakeList,
+  dragstart, dragover, dragleave, drop, dragend, remakeList, clearCompleted,
 } from './modifyList';
 import {
   updateTaskCompleted, inputCreateTask, tasks, updateTaskDescription,
@@ -87,18 +87,7 @@ const createList = (tasklist) => {
     li.textContent = 'Clear all completed';
     li.id = 'clear';
     li.addEventListener('click', () => {
-      const draggables = [...document.querySelectorAll('.draggable')];
-
-      const newList = draggables.filter((draggable) => draggable.getElementsByClassName('completed')[0].checked === false);
-
-      draggables.forEach((draggable) => ul.removeChild(draggable));
-
-      newList.forEach((item) => ul.appendChild(item));
-
-      remakeList();
-
-      const clear = document.getElementById('clear');
-      ul.appendChild(clear);
+      clearCompleted(ul);
     });
 
     return li;
