+++ src/modules/frontend/modifyList.js
@@ -21,6 +21,21 @@ const remakeList = () => {
   });
 };
 
+const clearCompleted = (ul) => {
+  const draggables = [...document.querySelectorAll('.draggable')];
+
+  const newList = draggables.filter((draggable) => draggable.getElementsByClassName('completed')[0].checked === false);
+
+  draggables.forEach((draggable) => ul.removeChild(draggable));
+
+  newList.forEach((item) => ul.appendChild(item));
+
+  remakeList();
+
+  const clear = document.getElementById('clear');
+  ul.appendChild(clear);
+};
+
 const dragstart = (element) => {
   element.classList.add('flying');
 };
@@ -50,5 +65,5 @@ const dragend = (element) => {
 };
 
 export {
-  dragstart, dragover, dragleave, drop, dragend, remakeList,
+  dragstart, dragover, dragleave, drop, dragend, remakeList, clearCompleted,
 };
